add evaluation module
